Basic skeleton of Patch class definition
Here I just wanted to show the basic structure of the Patch class, and how the methods and data types are organized in a typical C++ class. I have a constructor (allocates memory for an object of the class) and a deconstructor (deallocates memory for an object of the class) set up bare-bones (you actually don't need to type anything in them to run; we will customize the constructor soon though). The next step is writing in the Constructor function and creating "getter" and "setter" methods. One thing I'd like to know though is what exactly a "Patch" consists of and what it does, so I have an absolutely clear picture of how to approach the class definition. Thanks!